refactor: [sync general] reorganize file and component
