+++ src/components/dialog/DialogDisableSettings.js
@@ -7,7 +7,7 @@ import {
     ModalActions,
     ButtonStrip,
     Button,
-} from '@dhis2/ui-core'
+} from '@dhis2/ui'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
