+++ src/pages/General/DisableSettings.js
@@ -2,8 +2,8 @@ import React, { useState } from 'react'
 import { useDataMutation } from '@dhis2/app-runtime'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
-import { ButtonField } from '../../components/sections/general/form-sections'
-import DialogDisableSettings from '../../components/dialog/dialog-disable-settings'
+import { ButtonField } from '../../components/field'
+import DialogDisableSettings from '../../components/dialog/DialogDisableSettings'
 import { NAMESPACE } from '../../constants/data-store'
 import { useNavigation } from '../../utils/useNavigation'
